test(maintenance): give the admin-directory anchor tests their own file
test_gc_worktrees_real_git_healthy.py answers "does this worktree look healthy
and still hold unique work". The three refs/worktree tests answer a different
question: does an anchor that lives only in the admin directory keep a commit
alive, and does git agree. Moving them out leaves each file with one subject
and takes the healthy file from 458 lines to 359, with room for the reflog
cases that follow.

Pure move. The three test bodies and their helper are byte-identical to their
previous home; only the module docstring and the import list are new.

Copilot-Session: 3ab110a4-2e9e-4504-a731-74fa73874896
